refactor(nuxt): modularize module/config and document Nitro fixture
Split long module and config files into focused helpers, and clarify why
the strict playground keeps a Nitro /api/env route for auto-extend coverage.
