+++ apps/playgrounds/nuxt-strict/server/api/env.get.ts
@@ -1,3 +1,9 @@
+/**
+ * Nitro fixture for strict-layout auto-extend.
+ *
+ * Imports `~~/env/server` so the playground build exercises the Nitro alias /
+ * `__ARKENV_STRICT_LAYOUT__` wiring (Vite alone is not enough).
+ */
 import { env } from "~~/env/server";
 
 export default defineEventHandler(() => {
